Use module.exports to export default methods
Also export named exports as individual files
