+++ compile.js
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/no-unresolved
-module.exports = require('./dist/compile');
+module.exports = require('./dist/compile').default;
